Loggerprocess now exists and can communicate correctly

+++ Utility/Utilities/Include/Constants/LoggerConstants.hpp
@@ -9,7 +9,7 @@ namespace Doremi
         {
             const size_t FILEMAP_SIZE = 1024;
             const std::string FILEMAP_NAME = std::string("doremi_filemap");
-            const size_t FILEMAP_TIMEOUT = 10;
+            const double FILEMAP_TIMEOUT = 10;
             const size_t LONGEST_FUNCTION_NAME = 256;
             const size_t LONGEST_MESSAGE_NAME = 256;
             const size_t LONGEST_LINE_NAME = 256;
